backend/core/pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = ""
authors = ["nnnLik <rslmakhmudov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "==5.0.3"
pydantic-settings = "==2.2.1"
psycopg2-binary = "^2.9.9"
django-cors-headers = "^4.3.1"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
gunicorn = "^21.2.0"
pillow = "^10.3.0"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
